Moves pwf_evaluator out of the taxonomy header file and moves dependency on settings from pwf to pwf_evaluator
